# api-design/autogenerated/openapi.yaml
openapi: 3.0.0
info:
  title: TodoDoo API
  version: 0.0.0
...
paths:
...
  /api/todos:
    get:
      operationId: Todos_list
      description: List all todos for the authenticated user
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
    post:
      operationId: Todos_create
      description: Create a new todo
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
  /api/todos/generate-break-up-todos:
    post:
      operationId: Todos_generateBreakUpTodos
      description: Generate break-up todos across all todos
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
  /api/todos/{id}:
    get:
      operationId: Todos_get
      description: Get a specific todo by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
    patch:
      operationId: Todos_update
      description: Update a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
    delete:
      operationId: Todos_delete
      description: Delete a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
  /api/todos/{id}/reorder:
    post:
      operationId: Todos_reorder
      description: Reorder a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderTodoRequest'
  /api/todos/{id}/start-timer:
    post:
      operationId: Todos_startTimer
      description: Start a timer for a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
  /api/todos/{id}/stop-timer:
    post:
      operationId: Todos_stopTimer
      description: Stop a timer for a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
  /api/todos/{id}/time-elapsed-override:
    post:
      operationId: Todos_overrideElapsedTime
      description: Override elapsed time for a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OverrideElapsedTimeRequest'
...
security:
  - BearerAuth: []
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      description: Error response
...
    Importance:
      type: integer
      format: int32
      minimum: 0
      maximum: 5
      description: Importance level (0-5)
...
    OverrideElapsedTimeRequest:
      type: object
      required:
        - hours
        - minutes
      properties:
        hours:
          type: integer
          format: int32
          description: Hours component of elapsed time
        minutes:
          type: integer
          format: int32
          description: Minutes component of elapsed time
      description: Request body for overriding elapsed time
...
    ReorderTodoRequest:
      type: object
      required:
        - position
      properties:
        position:
          allOf:
            - $ref: '#/components/schemas/TodoPosition'
          description: New position for the todo
      description: Request body for reordering a todo
...
    Todo:
      type: object
      required:
        - id
        - task
        - status
        - date_created
        - date_updated
        - user_id
        - order
        - annoyingness
        - committed
        - elapsedTimeSeconds
        - atomic
        - infinitely_divisible
      properties:
        id:
          type: string
          description: Unique identifier for the todo
        task:
          type: string
          description: The task description
        details:
          type: object
          description: Additional details stored as JSON
        status:
          allOf:
            - $ref: '#/components/schemas/TodoStatus'
          description: Current status of the todo
        date_created:
          type: string
          format: date-time
          description: When the todo was created
        date_updated:
          type: string
          format: date-time
          description: When the todo was last updated
        user_id:
          type: string
          description: ID of the user who owns this todo
        order:
          type: string
          description: Order/position of the todo in the list
        priority:
          type: integer
          format: int32
          description: Integer priority position (0-based index in the ordered list)
        desirability:
          type: number
          format: float
          description: Calculated desirability score for weighted random selection
        importance:
          allOf:
            - $ref: '#/components/schemas/Importance'
          description: Importance level (0-5)
        time_estimate:
          allOf:
            - $ref: '#/components/schemas/TimeEstimate'
          description: Estimated time to complete
        annoyingness:
          allOf:
            - $ref: '#/components/schemas/Annoyingness'
          description: Annoyingness level (0-4)
        committed:
          type: boolean
          description: Whether the todo is committed
        elapsedTimeSeconds:
          type: integer
          format: int32
          description: Elapsed time in seconds for timer
        atomic:
          type: boolean
          description: Whether the todo is atomic (cannot be broken down)
        infinitely_divisible:
          type: boolean
          description: Whether the todo is infinitely divisible
      description: A todo item
...
    TodoStatus:
      type: string
      enum:
        - todo
        - complete
        - archived
        - waiting
      description: Status of a todo item
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: Bearer
servers:
  - url: http://localhost:8000
    description: Development server
    variables: {}